requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.9
pytest==7.4.0
flake8==6.1.0
